desktop-audio-player/requirements.txt
paho-mqtt>=2.0.0
httpx>=0.27
miniaudio>=1.61
